+++ src/main/java/com/epam/aidial/cfg/client/dto/ApplicationResourceDto.java
@@ -1,5 +1,6 @@
 package com.epam.aidial.cfg.client.dto;
 
+import com.epam.aidial.cfg.domain.value.LocalizedValue;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.AllArgsConstructor;
@@ -21,11 +22,11 @@ public class ApplicationResourceDto {
     private String endpoint;
     private String responsesEndpoint;
     private Map<String, DeploymentInterfaceResourceDto> interfaces;
-    private String displayName;
+    private LocalizedValue displayName;
     private String displayVersion;
     private String iconUrl;
-    private String description;
-    private String intro;
+    private LocalizedValue description;
+    private LocalizedValue intro;
     private String reference;
     private Boolean forwardAuthToken;
     private List<String> inputAttachmentTypes;
