+++ src/main/java/com/epam/aidial/cfg/mapper/ApplicationResourceMapper.java
@@ -1,6 +1,5 @@
 package com.epam.aidial.cfg.mapper;
 
-import com.epam.aidial.cfg.domain.mapper.LocalizedValueMapper;
 import com.epam.aidial.cfg.dto.ApplicationResourceDto;
 import com.epam.aidial.cfg.dto.ApplicationResourceNodeInfoDto;
 import com.epam.aidial.cfg.dto.ApplicationsEximDto;
@@ -16,7 +15,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.Named;
 
-@Mapper(componentModel = "spring", uses = {LocalizedValueMapper.class})
+@Mapper(componentModel = "spring")
 public interface ApplicationResourceMapper {
 
     @Mapping(target = "applicationTypeSchemaId", source = "source", qualifiedByName = "toSchemaIdString")
