+++ src/test/java/com/epam/aidial/cfg/service/CoreDeploymentServiceTest.java
@@ -9,6 +9,7 @@ import com.epam.aidial.cfg.client.dto.ModelLimitsDataDto;
 import com.epam.aidial.cfg.client.dto.ModelPricingDataDto;
 import com.epam.aidial.cfg.client.dto.ToolSetDataDto;
 import com.epam.aidial.cfg.client.mapper.DeploymentClientMapper;
+import com.epam.aidial.cfg.domain.value.LocalizedValue;
 import com.epam.aidial.cfg.model.ApplicationData;
 import com.epam.aidial.cfg.model.DeploymentType;
 import com.epam.aidial.cfg.model.InterfaceType;
@@ -105,7 +106,7 @@ class CoreDeploymentServiceTest {
         var modelDto = ModelDataDto.builder()
                 .id("gpt-4")
                 .object("dial-model")
-                .displayName("GPT-4")
+                .displayName(LocalizedValue.of("GPT-4"))
                 .capabilities(ModelCapabilitiesDataDto.builder().chatCompletion(true).build())
                 .limits(ModelLimitsDataDto.builder().maxPromptTokens(128000).build())
                 .pricing(ModelPricingDataDto.builder().unit("token").prompt("0.1").build())
@@ -113,14 +114,14 @@ class CoreDeploymentServiceTest {
         var applicationDto = ApplicationDataDto.builder()
                 .id("my-app")
                 .object("dial-application")
-                .displayName("My App")
+                .displayName(LocalizedValue.of("My App"))
                 .applicationTypeSchemaId("schema-1")
                 .applicationProperties(Map.of("key", "value"))
                 .build();
         var toolsetDto = ToolSetDataDto.builder()
                 .id("my-toolset")
                 .object("dial-toolset")
-                .displayName("My Toolset")
+                .displayName(LocalizedValue.of("My Toolset"))
                 .transport("streamable-http")
                 .allowedTools(List.of("tool-a"))
                 .build();
