+++ src/test/java/com/epam/aidial/cfg/service/ZipToolSetEximServiceTest.java
@@ -3,8 +3,6 @@ package com.epam.aidial.cfg.service;
 import com.epam.aidial.cfg.client.mapper.RouteMapperImpl;
 import com.epam.aidial.cfg.client.mapper.ToolSetClientMapperImpl;
 import com.epam.aidial.cfg.configuration.JsonMapperConfiguration;
-import com.epam.aidial.cfg.configuration.LocalizationProperties;
-import com.epam.aidial.cfg.domain.mapper.LocalizedValueMapper;
 import com.epam.aidial.cfg.domain.value.LocalizedValue;
 import com.epam.aidial.cfg.dto.ToolSetEximDto;
 import com.epam.aidial.cfg.dto.ToolSetsEximDto;
@@ -46,9 +44,7 @@ import static org.mockito.Mockito.when;
         ToolSetClientMapperImpl.class,
         ZipToolSetEximService.class,
         ResourceImportValidator.class,
-        RouteMapperImpl.class,
-        LocalizedValueMapper.class,
-        LocalizationProperties.class
+        RouteMapperImpl.class
 })
 class ZipToolSetEximServiceTest {
